guangzhou-yunnan-sales/backend/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
mysqlclient==2.2.4
Pillow==10.4.0
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
statsmodels==0.14.4
python-dateutil==2.9.0
